Bugs: First click not detected and text selections
1. If the user clicks a pixel and then hover the rest of the pixels,
the first one will not change its color. This is because of the DOM
"click" event that fires and the user releases the click. The solution
was to change the event "click" to "mousedown"

2. When clicking and hovering over the canvas, some labels and the
text input will be selected, interrupting the user to keep painting
on the canvas. The solution was to apply all elements the
"user-select: none" rule to avoid this problem